requirements.txt
pystray 
Pillow